+++ polonius-engine/src/output/datafrog_opt.rs
@@ -41,7 +41,6 @@ pub(super) fn compute<Region: Atom, Loan: Atom, Point: Atom, Variable: Atom, Mov
         all_facts.var_drops_region,
         var_maybe_initialized_on_exit,
         &all_facts.cfg_edge,
-        all_facts.region_live_at,
         all_facts.universal_region,
         &mut result,
     );