+++ src/program.rs
@@ -18,7 +18,6 @@ struct Facts {
     cfg_edge: BTreeSet<(Point, Point)>,
     killed: BTreeSet<(Loan, Point)>,
     outlives: BTreeSet<(Region, Region, Point)>,
-    region_live_at: BTreeSet<(Region, Point)>,
     invalidates: BTreeSet<(Point, Loan)>,
     var_defined: BTreeSet<(Variable, Point)>,
     var_used: BTreeSet<(Variable, Point)>,
@@ -39,7 +38,6 @@ impl From<Facts> for AllFacts {
             cfg_edge: facts.cfg_edge.into_iter().collect(),
             killed: facts.killed.into_iter().collect(),
             outlives: facts.outlives.into_iter().collect(),
-            region_live_at: facts.region_live_at.into_iter().collect(),
             invalidates: facts.invalidates.into_iter().collect(),
             var_defined: facts.var_defined.into_iter().collect(),
             var_used: facts.var_used.into_iter().collect(),
@@ -145,26 +143,13 @@ pub(crate) fn parse_from_program(
             // the most common statement effects: mid point effects
             for effect in &statement.effects {
                 match effect {
-                    Effect::Use { ref regions } => {
-                        // Uses.
-                        // TODO: Incomplete. We should eventually compute liveness
-                        // in order to emit `region_live_at` facts at all correct computed points,
-                        // and not just at the manually specified statements' Start point.
-                        //
-                        // facts: region_live_at(Region, Point)
-                        // region_live_at: a `use` emits a `region_live_at` the Start point
-                        facts
-                            .region_live_at
-                            .extend(regions.into_iter().map(|region| {
-                                let region = tables.regions.intern(region);
-                                (region, start)
-                            }));
-                    }
-
+                    // TODO: once the parser is revamped for liveness etc, make
+                    // sure to catch the new inputs here!
                     Effect::Fact(ref fact) => {
                         // Manually specified facts
                         emit_fact(&mut facts, fact, mid, tables)
-                    }
+                    },
+                    _ => {}
                 };
             }
 
@@ -217,14 +202,6 @@ fn emit_fact(facts: &mut Facts, fact: &Fact, point: Point, tables: &mut Interner
             facts.invalidates.insert((point, loan));
         }
 
-        // facts: region_live_at(Region, Point)
-        Fact::RegionLiveAt { ref region } => {
-            let region = tables.regions.intern(region);
-            // region_live_at: a region can be manually set live on both Start and Mid points
-            // but will mostly be computed and emitted automatically
-            facts.region_live_at.insert((region, point));
-        }
-
         // facts: var_defined(V, P)
         Fact::DefineVariable { ref variable } => {
             // var_defined: a variable is overwritten here
@@ -237,7 +214,9 @@ fn emit_fact(facts: &mut Facts, fact: &Fact, point: Point, tables: &mut Interner
             // var_used: a variable is used here
             let variable = tables.variables.intern(variable);
             facts.var_used.insert((variable, point));
-        }
+        },
+
+        _ => {},
     };
 }
 
@@ -303,30 +282,6 @@ mod tests {
 
             assert_eq!(point, "\"Start(B0[1])\"");
             assert_eq!(loan, "L1");
-        }
-
-        // TODO: incomplete until either all the `region_live_at` are computed with liveness,
-        // or they are emitted manually at Start points.
-        // facts: region_live_at
-        assert_eq!(facts.region_live_at.len(), 3);
-        {
-            let region = tables.regions.untern(facts.region_live_at[0].0);
-            let point = tables.points.untern(facts.region_live_at[0].1);
-
-            assert_eq!(region, "'a");
-            assert_eq!(point, "\"Start(B1[0])\"");
-
-            let region = tables.regions.untern(facts.region_live_at[1].0);
-            let point = tables.points.untern(facts.region_live_at[1].1);
-
-            assert_eq!(region, "'b");
-            assert_eq!(point, "\"Start(B1[0])\"");
-
-            let region = tables.regions.untern(facts.region_live_at[2].0);
-            let point = tables.points.untern(facts.region_live_at[2].1);
-
-            assert_eq!(region, "'d");
-            assert_eq!(point, "\"Start(B0[1])\"");
         }
 
         assert_eq!(facts.outlives.len(), 1);